Fix: Class "Dashboard" has 30 methods, which is greater than 20 authorized. Split it into smaller classes.
